requirements.txt
customtkinter>=5.2.0
paho-mqtt>=2.0.0
requests>=2.31.0
Pillow>=10.0.0
pystray>=0.19.0
plyer>=2.1.0
pygame>=2.5.0
